s2l circles

+++ generators/g_circles.py
@@ -94,7 +94,7 @@ class g_circles():
         if self.channel >= 0:
             current_volume = float(str(self.sound_values.buf[self.channel*8:self.channel*8+8],'utf-8'))
             self.number = int(10 * current_volume)
-            self.soundsize = np.clip(3 * current_volume, 0, 3)
+            self.soundsize =int(np.clip(3 * current_volume, 0, 3))
 
         for i in range(self.number):
             # check if S2L is activated
